feat(cli): add initial CLI support for object manifests (#97)
* feat(cli): add initial CLI support for object manifests

* Apply automatic documentation changes

* Apply automatic documentation changes

* feat(objects): add more options for click for manifests

* Apply automatic documentation changes

* Update gen3/tools/indexing/verify_manifest.py


* Apply automatic documentation changes

* Apply automatic documentation changes

* fix(jenkins): update to branch of jenkins lib to test fix

* Apply automatic documentation changes

* Update Jenkinsfile

* feat(delete): add support for deleting GUIDs

* Apply automatic documentation changes

* fix(download_manifest): make download manifest deterministic, sort array-like columns

* fix(python): use python3

* Apply automatic documentation changes

* chore(deps): install fastavro

* Apply automatic documentation changes

* Update gen3/cli/objects.py


* Update gen3/cli/objects.py


* Apply automatic documentation changes

* Update gen3/cli/objects.py


* Apply automatic documentation changes

* fix(review): adjust based on review, cleanup logging, more performance parsing of GUID, remove unnecessary var

* Apply automatic documentation changes

* feat(objects): move delete guids out of CLI into SDK so it can be used by both

* Apply automatic documentation changes

+++ gen3/index.py
@@ -740,7 +740,8 @@ class Gen3Index:
 
         """
         rec = self.client.get(guid)
-        rec.delete()
+        if rec:
+            rec.delete()
         return rec
 
     ### Query Requests
